Bot will periodically initiate a random coinflip game for user to accept for free

+++ handlers/buttons/misc/coinflip.js
@@ -17,19 +17,20 @@ async function initCoinflip(client, guild, channel, gameCode) {
     await channel.send({ content: `<:botconfirm:845719660812435496> <@${playerTwo}> has accepted <@${playerOne}>'s wager of **${wagerAmount / 2}** tokens. Good luck!` }).catch(err => console.error(err));
     // Get a random number, 1 = playerOne, 2 = playerTwo
     const pickWinner = randomNum(1, 2);
+    let checkWinnersTokens;
     if (pickWinner === 1) {
-        // Get the winners current tokens count and add the wagered tokens
-        const checkWinnersTokens = await dbFindOne(tokensSchema, { userId: playerOne });
-        await dbUpdateOne(tokensSchema, { userId: playerOne }, { tokens: checkWinnersTokens.tokens + wagerAmount });
-        // Notify players of the result
-        // await channel.send({ content: `<:dollar_coin:1034396609276022854> <@${playerOne}> won the wager of **${wagerAmount}** tokens against <@${playerTwo}>!` }).catch(err => console.error(err));
+        if (playerOne !== client.user.id) {
+            // Get the winners current tokens count and add the wagered tokens
+            checkWinnersTokens = await dbFindOne(tokensSchema, { userId: playerOne });
+            await dbUpdateOne(tokensSchema, { userId: playerOne }, { tokens: checkWinnersTokens.tokens + wagerAmount });
+        }
         await createCanvas(client, guild, channel, playerOne, playerTwo, playerOne, wagerAmount, gameCode, checkWinnersTokens);
     } else {
-        // Get the winners current tokens count and add the wagered tokens
-        const checkWinnersTokens = await dbFindOne(tokensSchema, { userId: playerTwo });
-        await dbUpdateOne(tokensSchema, { userId: playerTwo }, { tokens: checkWinnersTokens.tokens + wagerAmount });
-        // Notify players of the result
-        // await channel.send({ content: `<:dollar_coin:1034396609276022854> <@${playerTwo}> won the wager of **${wagerAmount}** tokens against <@${playerOne}>!` }).catch(err => console.error(err));
+        if (playerTwo !== client.user.id) {
+            // Get the winners current tokens count and add the wagered tokens
+            checkWinnersTokens = await dbFindOne(tokensSchema, { userId: playerTwo });
+            await dbUpdateOne(tokensSchema, { userId: playerTwo }, { tokens: checkWinnersTokens.tokens + wagerAmount });
+        }
         await createCanvas(client, guild, channel, playerOne, playerTwo, playerTwo, wagerAmount, gameCode, checkWinnersTokens);
     }
     // Remove the game entry from the database
@@ -76,7 +77,7 @@ async function createCanvas(client, guild, channel, playerOne, playerTwo, winner
     await channel.createWebhook({ name: client.user.username, avatar: client.user.avatarURL({ format: 'png', size: 256 }) })
         .then(webhook => {
             webhook.send({
-                content: `**Game:** <@${playerOne}> vs. <@${playerTwo}> \n**Winnings:** ${wagerAmount * 2} tokens \n**Code:** ${gameCode} \n\n*Create your own wager with </coinflip create:1064426324690751528>*`,
+                content: `**Game:** <@${playerOne}> vs. <@${playerTwo}> \n**Winnings:** ${wagerAmount} tokens \n**Code:** ${gameCode} \n\n*Create your own wager with </coinflip create:1064426324690751528>*`,
                 files: [attachment]
             }).catch(err => console.error(`${path.basename(__filename)} There was a problem sending a webhook message: `, err))
                 .then(() => {
@@ -86,12 +87,14 @@ async function createCanvas(client, guild, channel, playerOne, playerTwo, winner
                 });
         }).catch(err => console.error(`${path.basename(__filename)} There was a problem sending a webhook: `, err));
     // Log winner's increase in tokens
-    tokenLog.send({
-        content: `${process.env.TOKENS_UP} <@${winnerId}> gained **${wagerAmount / 2}** tokens from a coinflip, they now have **${checkWinnersTokens.tokens + wagerAmount}** tokens`,
-        allowedMentions: {
-            parse: []
-        }
-    }).catch(err => console.error(`${path.basename(__filename)} There was a problem sending a message: `, err));
+    if (winnerId !== client.user.id) {
+        tokenLog.send({
+            content: `${process.env.TOKENS_UP} <@${winnerId}> gained **${wagerAmount / 2}** tokens from a coinflip, they now have **${checkWinnersTokens.tokens + wagerAmount}** tokens`,
+            allowedMentions: {
+                parse: []
+            }
+        }).catch(err => console.error(`${path.basename(__filename)} There was a problem sending a message: `, err));
+    }
 }
 
 function randomNum(min, max) {
@@ -109,7 +112,7 @@ module.exports = async (interaction) => {
     const gameCode = customId.split('-')[1];
 
     interaction?.fetchReply('@original').then(async reply => {
-        if (reply.content.toLowerCase().includes('head-to-head')) {
+        if (reply.content.toLowerCase().includes('head-to-head') || reply.content.toLowerCase().includes('play for free')) {
             // Get the game data from the database
             const coinflipGame = await dbFindOne(coinflipSchema, { code: gameCode });
             // Don't let a user join their own game
@@ -117,33 +120,35 @@ module.exports = async (interaction) => {
             // Check if the owner of the coinflip is still in the server, if not, delete their game
             const checkIfUserExists = await guild.members.fetch(coinflipGame.playerOne).catch(err => console.error(err));
             if (!checkIfUserExists) {
-                reply.edit({ content: `${reply.content.split('.')[0]}. The wager is no longer valid`, components: [] }).catch(err => console.error(err));
+                reply.edit({ content: `${reply.content.split('.')[0]}. This wager is no longer valid`, components: [] }).catch(err => console.error(err));
                 return dbDeleteOne(coinflipSchema, { code: gameCode });
             }
             // If the coinflip is already in progress
             if (coinflipGame?.inProgress === true) return;
             // Check if the user has enough tokens to wager
-            const checkUserTokens = await dbFindOne(tokensSchema, { userId: member.id });
-            // If the user doesn't have a tokens database entry
-            if (!checkUserTokens)
-                return channel.send({ content: `${process.env.BOT_DENY} ${member} You don't have enough tokens to join this wager, start earning tokens by chatting in the server` })
-                    .catch(err => console.error(err))
-                    .then(message => {
-                        setTimeout(() => {
-                            message.delete().catch(err => console.error(err));
-                        }, 7000);
-                    });
-            // If the user doesn't have enough tokens to join the wager
-            if (checkUserTokens.tokens < coinflipGame.amount)
-                return channel.send({ content: `${process.env.BOT_DENY} ${member} You don't have enough tokens to join this wager, your current tokens balance is ${checkUserTokens.tokens}` })
-                    .catch(err => console.error(err))
-                    .then(message => {
-                        setTimeout(() => {
-                            message.delete().catch(err => console.error(err));
-                        }, 7000);
-                    });
-            // Decuct the wagered tokens now so the user can't spend them before the game is finished
-            await dbUpdateOne(tokensSchema, { userId: member.id }, { tokens: checkUserTokens.tokens - coinflipGame.amount });
+            if (coinflipGame.playerOne !== client.user.id) {
+                const checkUserTokens = await dbFindOne(tokensSchema, { userId: member.id });
+                // If the user doesn't have a tokens database entry
+                if (!checkUserTokens)
+                    return channel.send({ content: `${process.env.BOT_DENY} ${member} You don't have enough tokens to join this wager, start earning tokens by chatting in the server` })
+                        .catch(err => console.error(err))
+                        .then(message => {
+                            setTimeout(() => {
+                                message.delete().catch(err => console.error(err));
+                            }, 7000);
+                        });
+                // If the user doesn't have enough tokens to join the wager
+                if (checkUserTokens.tokens < coinflipGame.amount)
+                    return channel.send({ content: `${process.env.BOT_DENY} ${member} You don't have enough tokens to join this wager, your current tokens balance is ${checkUserTokens.tokens}` })
+                        .catch(err => console.error(err))
+                        .then(message => {
+                            setTimeout(() => {
+                                message.delete().catch(err => console.error(err));
+                            }, 7000);
+                        });
+                // Decuct the wagered tokens now so the user can't spend them before the game is finished
+                await dbUpdateOne(tokensSchema, { userId: member.id }, { tokens: checkUserTokens.tokens - coinflipGame.amount });
+            }
             // If the game was found, set playerTwo as the challenger and start the game
             await dbUpdateOne(coinflipSchema, { code: gameCode }, { playerTwo: member.id, inProgress: true });
             // Mark this wager as accepted
